+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Release Notes for Gatsby Helper
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where queries that would not be defined by the GraphQL schema were exposed to Gatsby, anyway.
+
 ## 1.0.0-beta.1 - 2020-11-03
 
 ### Added