+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes for Gatsby Helper
 
-## Unreleased
+## 1.0.0-beta.2 - 2020-12-01
 
 > {note} You will need to ensure that your “Preview Webhook URL” setting is set correctly after updating.